Projects/NLP-Assistant/backend/requirements.txt
torch==1.10.0
torchaudio==0.10.0
numpy==1.21.4
scipy==1.7.3
transformers==4.15.0
fastapi==0.70.0
uvicorn==0.15.0
python-multipart==0.0.5
sounddevice==0.4.4
soundfile==0.10.3.post1
webrtcvad==2.0.10
pyaudio==0.2.11
pydub==0.25.1
librosa==0.8.1
nltk==3.6.5
spacy==3.2.0
pyttsx3==2.90
websockets==10.1
pytest==6.2.5
python-dotenv==0.19.2
